add vision in 4 dirs and add an input layer

+++ Vector.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass(unsafe_hash=False)
 class Vector:
-    x: float
-    y: float
+    x: int
+    y: int
